+++ client/flags/flags.go
@@ -83,11 +83,12 @@ const (
 	FlagReverse          = "reverse"
 	FlagTip              = "tip"
 	FlagAux              = "aux"
+	FlagInitHeight       = "initial-height"
 	// FlagOutput is the flag to set the output format.
 	// This differs from FlagOutputDocument that is used to set the output file.
 	FlagOutput = cmtcli.OutputFlag
 
-	// CometBFT logging flags
+	// Logging flags
 	FlagLogLevel  = "log_level"
 	FlagLogFormat = "log_format"
 )