+++ x/gov/client/cli/query.go
@@ -38,6 +38,7 @@ func GetQueryCmd() *cobra.Command {
 		GetCmdQueryDeposit(),
 		GetCmdQueryDeposits(),
 		GetCmdQueryTally(),
+		GetCmdConstitution(),
 	)
 
 	return govQueryCmd
@@ -651,4 +652,26 @@ $ %s query gov proposer 1
 	flags.AddQueryFlagsToCmd(cmd)
 
 	return cmd
+}
+
+func GetCmdConstitution() *cobra.Command {
+	return &cobra.Command{
+		Use:   "constitution",
+		Short: "Get the constitution",
+		Args:  cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+			queryClient := v1.NewQueryClient(clientCtx)
+
+			resp, err := queryClient.Constitution(cmd.Context(), &v1.QueryConstitutionRequest{})
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(resp)
+		},
+	}
 }