+++ x/gov/migrations/v5/store.go
@@ -24,6 +24,8 @@ func MigrateStore(ctx sdk.Context, storeKey storetypes.StoreKey, cdc codec.Binar
 	params.ExpeditedMinDeposit = defaultParams.ExpeditedMinDeposit
 	params.ExpeditedVotingPeriod = defaultParams.ExpeditedVotingPeriod
 	params.ExpeditedThreshold = defaultParams.ExpeditedThreshold
+	params.ProposalCancelRatio = defaultParams.ProposalCancelRatio
+	params.ProposalCancelDest = defaultParams.ProposalCancelDest
 
 	bz, err := cdc.Marshal(&params)
 	if err != nil {