+++ x/tx/textual/repeated_test.go
@@ -12,8 +12,8 @@ import (
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protoreflect"
 
+	"cosmossdk.io/x/tx/internal/testpb"
 	"cosmossdk.io/x/tx/textual"
-	"cosmossdk.io/x/tx/textual/internal/testpb"
 )
 
 type repeatedJsonTest struct {